requirements.txt
asgiref==3.11.1
Django==6.0.2
et_xmlfile==2.0.0
mysqlclient==2.2.7
openpyxl==3.1.5
pillow==12.1.0
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
